+++ plugins/frontend/backstage-plugin-cloudsmith/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @roadiehq/backstage-plugin-cloudsmith
 
+## 0.1.3
+
+### Patch Changes
+
+- 054d585b: Bump plugin versions to be compatible by backstage 1.9.1
+
 ## 0.1.2
 
 ### Patch Changes